Remove passes from CustomMaterial frontend API
For example:

CustomMaterial {
    property real time: 0.0
    property real amplitude: 5.0

    vertexShader: "example.vert"
    fragmentShader: "example.frag"
}

i.e. a lot like ShaderEffect. The properties are URLs, with the usual
rules (file or qrc or relative).

Also converts all built-in materials to the new style.

The multi-pass FrostedGlassMaterial is gone, and the name is reused
for the single pass version.

The glass materials lack blending and refraction for now. Blending may
be reintroduced later, while refraction depends on screen reading
support.

Change-Id: Ib75fea549d2cec204fd11e76c0b662b699f13370
Reviewed-by: Andy Nichols <andy.nichols@qt.io>