Remove unnecessary "name" property from QQuick3DGeometry
The name property comes from the pattern used in Qt 3D Studio. This is
unnecessary though from a user perspective because in QML any item can
be referenced by its id.  On the backend there just needs to be a
differnt way to track custom geometry items in the renderbuffer manager
than a user defined string.  It is trivial to just track the geometry by
the rendergeometry scenegraph object which defines the geometry itself.

Change-Id: Icc74b7dc6335d815cc1fd409d0a02f00da443225
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>