Decouple custom materials from QSSGCommand
The command system is only used by effects now. There are a number of
commands (and related frontend APIs) that are never used in practice.
(either because they do not make sense for effects or because it was not
implemented for effects) Remove these.

Render modes such as blending and cull mode are important for custom materials,
these will be brought back in a simpler form in a follow up patch.

Change-Id: I17c01d239868951a1048a314111804134b91ebe8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>