Enable specifying blend factors and cull mode in custom materials
Also remove the hasRefraction property: this is effectively unused
internally, it is merely an alias to hasTransparency in practice.
(this does not affect the Refraction shader key, that's independent of
this)

The cull mode support is inherited from the base class (Material) so
setting that property should just work. (defaults to BackFaceCullign)

Blending is now handled by two properties (source and destination factor).
These will often be set together with hasTransparency. By default
objects with a custom material are treated as opaque and there will be
no blending. When hasTransparency is enabled, and the blend factors are
set to a valid value, the object will move to the non-opaque object list
and blending will get enabled.

The customshader example is enhanced to demonstrate this. When the alpha
slider is at the default 1.0, there is no blending and the object is
treated as opaque. When moving it to a lower value, blending will
get enabled and the object is part of the transparent object list.

All properties are freely changeable at runtime. This previously was not
true, but now hasTransparency, blend factors, etc. can change and the new
values are picked up automatically by the backend.

Change-Id: I9be8a4493d3f0744abd5511303d3202ee7a247f1
Reviewed-by: Andy Nichols <andy.nichols@qt.io>