Use updatePropertyListenser pattern in QQuick3DModel
For referenced properties like skeleton and geometry it is important to
listen for the destruction of the object so that that we don't try and
use an object that has been deleted on the backend.

The updatePropertyListener function will create a connection that
listens for the destruction of the referenced object, and will make sure
that the property is unset if the object is deleted.

Both Skeleton and Geometry references were doing partial checking for
this, but this changes makes it where everything uses the same pattern
that is used elsewhere. It also simplifies the ref and deref from the
scene manager.

Change-Id: Id3adfcc4fa334bb832f64b699d01fdede1c6690d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>