Add support for vertex processor in shaded custom materials
The shaded materials' custom vertex shader is pretty similar to
unshaded, except that the additional stuff used for e.g. lighting
(normals etc.) is calculated normally after calling it.

This means that the function has full control over gl_Position, and can
generate varying data to be passed to the fragment shader, based on NORMAL,
NORMAL_MATRIX, CAMERA_POSITION, MODEL_MATRIX, etc.

The depth prepass, shadow pass, etc. are not using the materials' vertex shaders
at all. This means that any modifications to the vertex data in the vertex shaders
are not reflected in the depth texture or shadow maps. This probably affects other
things as well, such as skinning. This is to be handled separately since it is
likely a feature on its own.

Change-Id: I45b46df0f5c23fb7b1eee72cfad549297a013654
Reviewed-by: Andy Nichols <andy.nichols@qt.io>