Add support for specular light processors
Also rename LIGHT_DIRECTION to TO_LIGHT_DIR, in an attempt to indicate
that it is the direction vector pointing towards the light, not the
other way round. Sadly the shader generator source code is barely
readable, but we should try making at least the user facing names clear.

Drop BASE_COLOR as well, it does not seem useful for now.

Change-Id: Ifc93ae4559df5c2d0493d8022d81e55c658df3f0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>