Add support for custom textures from data
Previously the only way to generate dynamic textures at runtime was to
use QML items. That is limited in what formats can be used (RGB8 and
RGBA8) so this adds an API to allow textures to be defined in C++ and
used by the backend.

This change introduces a new public C++ API in the form of
QQuick3DTextureData which can be subclassed by users to define
procedural textures at runtime by defining the data, format, and size of
the texture. This QQuick3DTextureData subclass can then be exposed to
the QML runtime and used as input to the textureData property of the
Texture component, thus enabling a third way to define the source of
Texture.

This change also moves the format property of the Texture component to
QQuick3DTextureData class, as this is where it makes more sense.  For
Texture components that uses either the source or sourceItem property
the automatic detection of format based on the source file or QML item
should be enough.

Change-Id: I8454f8834f639bc639c51a75a4190c49f88fefe7
Reviewed-by: Christian Strømme <christian.stromme@qt.io>