Repurpose custom fragment main in shaded custom materials
Now it is called early on in main(), and its task is to set METALNESS,
ROUGHNESS, SPECULAR_AMOUNT, and FRESNEL_POWER, if relevant.

This way it is possible to achieve the same what a PrincipledMaterial
would do. For example, instead of a SPECULAR_LIGHT function, the shader
can choose to add a MAIN instead and set METALNESS or SPECULAR_AMOUNT
in there, thus using the built-in specular calculations. Same goes for
light probes, where this is even more important, since there is no
processor function for probes, rather, the parameters are expected to be
set in MAIN().

In practice there are two cases:

- Dielectric: Set SPECULAR_AMOUNT to > 0. Optionally set ROUGHNESS to > 0.
- Non-dielectric: Set METALNESS to > 0. Set FRESNEL_POWER to > 0.
  Optionally set ROUGHNESS to > 0.

Change-Id: I16d30898af6ec77cf5c44ab285cb651e5b951000
Reviewed-by: Andy Nichols <andy.nichols@qt.io>