Add qt_ prefix for internal varyings
And expose some of them to custom shaders. These are not essential, a
custom vertex and fragment shader can manage without them by using their
own VARYINGs, but it's wasteful to calculate and pass the same thing
twice. So e.g. calculating your own normal and passing it to the
fragment shader one can just write VAR_NORMAL in the fragment shader
instead.

Also fixes vertex color support with custom materials. The vertex color
should always be present (and accessible/modifiable via COLOR in the
custom main), and it needs to default to vec4(1.0) when there are no
colors in the mesh.

Change-Id: Ifd7ae344ef68a84a8212fc5d14e417ced2369118
Reviewed-by: Andy Nichols <andy.nichols@qt.io>