Effect system: do not crash on missing image for texture property
The texture properties will be referred to by the shader, and if the
samplers are not bound to real texture data, some graphics APis,
e.g. vulkan, will crash. Avoid this by binding to a dummy texture if
the real one is missing. This is similar to what the custom material
system does.

Task-number: QTBUG-83177
Change-Id: Id11cefa773dc9d2c9dec3a29e1141c644ce3f1c7
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>