Use index of refraction value for something
Using the opacity in the fresnel and other calculations is surely wrong.
This seems to have originated from an earlier patch that replaced
material_specular.w with material_properties.w which have nothing to do
with each other (the former is the IOR, the latter is the object opacity,
also aliased in objectOpacity).

At least now one can see something in the principledmaterial example
when switching to non-metallic (specular amount > 0) and changing the
IOR value with the slider.

Change-Id: I8371d7a420e73604808fe7da4eb565b9c0967ae9
Reviewed-by: Andy Nichols <andy.nichols@qt.io>