Remove first level shader cache
QSSGRenderer should not need to keep a cache of the generated material
shaders. QSSGShaderCache has its own. What's more, the renderer's does
not include the ShaderFeatureSetList in the cache key which is fairly
incorrect and is a problem once one attempts to generate shaders for the
same mesh and so same material with different feature sets.

Change-Id: Ic2473d639b4e161cc68d3b8ed28cfb71710e8916
Reviewed-by: Christian Strømme <christian.stromme@qt.io>