Make sure we check Models with custom geometry when picking
Before the picking did not consider Models that uses geometry instead of
static meshes.  Now we make sure to at least check for ray casts against
the bounding volume set for the custom geometry.

This is for 5.15 only since there is a better way to do this in that is
not compatible with 5.15.

Fixes: QTBUG-84157
Change-Id: I08d0de80131d01dd97a6a0fb35da77ca7d97855e
Reviewed-by: Christian Strømme <christian.stromme@qt.io>