Add assimp import option to preserve pivot hierarchies
And turn it off by default since this can lead to very deep QML files
that are slow to load and edit.  The only time these are necessary is
with certain types of animations. So only turn this on when needed.

Fixes: QTBUG-85147
Change-Id: I98b9c2e25b9811fe729afc30b37d2e626fc71a05
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 171406ff4509ac49d1889a2404a302db39aa9383)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>