Expose depth, ao textures for custom material shaders
...and prepare for exposing a SCREEN_TEXTURE as well.

Exercise this in an unshaded material in the ssao manual test (only AO_TEXTURE
is used there, in practice DEPTH_TEXTURE is probably more useful. SCREEN_TEXTURE
is not implemented yet) and in a Lancelot test case.

For DEPTH_TEXTURE (and CAMERA_PROPERTIES for that matter) add a somewhat
dubious Lancelot test case.

Change-Id: I0661f1841b95c2a0a169d46991e3605b36259c4c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>