Do not register CustomMaterial shaders in updateSpatialNode
When we register CustomMaterial shaders in updateSpatialNode it requires
that we get a handle the the renderInterface which is unique to each
Window. That means we end up in a situtation where only one Window will
have the necessary shader registration if we share CustomMaterial nodes
between Windows.

Instead we store the necessary state information of the CustomMaterial
shaders in the CustomMaterial backend node.  That means that no which
renderer references the CustomMaterial it has all the information
necessary to create the shader.

The downside is that this means that there is additional overhead when
using CustomMaterials because each time we prep a layer for rendering we
need to register the CustomMaterial shaders being used with the dynamic
object system.  It will only do it once per shader, but not as ideal as
doing it once each time the shader data changes.

This makes it possible to share CustomMaterials between windows though.

This patch is made directly for 5.15 because CustomMaterials work very
differntly in 6.0 and this only makes sense in 5.15.

Fixes: QTBUG-83889
Change-Id: I57c72af12e55581874f66b8bd190ccbc918caf4d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>