Remove all unused textures from MaterialLibrary
We only released a sub-set CustomMaterials that we ported from Qt 3D
Studio to QtQuick3D's material library, but did not clean up the
textures they were using. This makes the MaterialLibrary larger than it
needs to be. Previously when building statically on Windows 32 bit it was
so large the QRC crashed. This can be avoided by just not shiping files
that are not being used.

This is only being done for 5.15 because the Material Library in 6.0 is
being change significantly enough to make backporting difficult.

Fixes: QTBUG-84188
Change-Id: I974282dfb0adb98c4b931abeaad895a16705939f
Reviewed-by: Christian Strømme <christian.stromme@qt.io>