Make sure to delete the Vertex Array when deleting a Input Assembler
A new VAO is created each time we create a new Input Assembler, which
itself is not an issue. However by not deleting the previous VAO that
referenced the vertex buffers associated with it, the graphics driver
does not release that buffer data, leading to huge leaks of GPU memory.

So now we make sure to call glDeleteVertexArrays for the backends that
use VertexArrays.

This change is for 5.15 only because 6.0 uses RHI which does not have
this issue.

Fixes: QTBUG-83808
Change-Id: I30a660c11f3aa89d9b5d0a9d4fdb8d1108d20702
Reviewed-by: Christian Strømme <christian.stromme@qt.io>