Expose all vertex inputs in the depth pass
...while starting to prepare the shadow map changes.

A custom vertex shader may want to transform the position based on some
of the vertex inputs (e.g. the normals). Such a vertex shader could not
generate proper output in the depth pass because it only saw dummy
normal and other values. Now instead we stop using the attr_pos-only
iaDepth and rather switch to the full input layout that the main
rendering pass uses.

Change-Id: Ibeeac3cb6899bcaf03ee678e3143cc925fc628bd
Reviewed-by: Andy Nichols <andy.nichols@qt.io>