Ship with optimized SPIR-V for the built-in shaders
Also includes stripping the shaders properly. While we do not have debug
information, we do have OpName and similar that describe the variable and
block member names. This is unnecessary in the final SPIR-V binary, and further
reduces the size somewhat.

To create debuggable shaders (for SPIR-V and D3D) with full source information,
one needs to regenerate the relevant ones with -g instead of -O.

Change-Id: I9a4dc0323c5478d7d0bd04848a8b56c4b69dcbbe
Reviewed-by: Andy Nichols <andy.nichols@qt.io>