Render shadow maps through the material shader generators
In the custommaterials example and in Lancelot tests such as
customdiffuse and customdiffusespecular this will fix the shadows
for models that have a material with a custom vertex shader
displacing the vertices.

The recently enhanced lights example can also showcase this: by
default nothing changes, even though internally the way shadow maps
are generated is fairly different. When enabling the custom material
on the teapot, it should now render correctly displaced shadows
for all light types.

Change-Id: I3c587d945ecca37fba33410a19213edda808595a
Reviewed-by: Christian Strømme <christian.stromme@qt.io>