Make lights example more useful
Restore the custom material option. Now a custom vertex shader that
can be used to verify that shadows look like they should even when
the vertices are displaced.

Add an option to stop the teapot rotation. Particularly useful in
combination with the above since it allows seeing better what is
going on.

Default to only the directional light enabled. Enabling all lights with
shadows brings any non-highend desktop system to its knees, which is not
ideal with an example. Rather, have a single directional light on by
default.

Change-Id: I40ea39ab32ec7651b55caad3a264db7660fffe05
Reviewed-by: Andy Nichols <andy.nichols@qt.io>