Drop the item2D check from depth pass
Not much point in disabling the depth texture (or the depth prepass)
when there's a 2D item in the scene. Just pretend those are transparent
and so not part of the depth buffer. Otherwise we cannot use SSAO or
custom materials relying on DEPTH_TEXTURE whenever having 2D items
inline.

This is from the times when Item2D was rendering textured quads,
based on the thinking that those quads should be a full part of the
3D scene, reflected in any 3D feature. This is not the case in practice.

Change-Id: I2d1af90d75b48781509ed116297cde011e50fd7b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>