Remove hasTransparency from CustomMaterial
No real need for this. Setting the blending, using SCREEN_TEXTURE, or
having a global opacity < 1 for the model will all lead to making the
object part of the semi-transparent list. No need for a dedicated flag,
especially since getting proper alpha blending will always involve
setting the blend modes to something other than the default NoBlend.
This simplifies the public facing API a bit, and is also simpler
internally.

Change-Id: Ifa653f0cccf9f3709ca3c1cbf76ba91e554b83ae
Reviewed-by: Andy Nichols <andy.nichols@qt.io>