Update QML versioning and type registration
Use the new API for registering QML types. Also update to the new
versioning standard: Starting with Qt 6.0, the QML import version
is the same as the Qt version. Finally, use versionless QML imports
in the library files.

This also moves the base QML types for Effects and Materials into
the QtQuick3D import. Since there are no custom materials in the
library currently, the QtQuick3D.Materials import is no longer
available, and will give an error. The examples and lancelot tests
are updated to fix this.

[ChangeLog] The QtQuick3D.Materials import is no longer available.
CustomMaterial is now part of the base QtQuick3D import.

Fixes: QTBUG-86065
Fixes: QTBUG-86064
Change-Id: Ib1fd4f487e8b882ceaddcd82be19b5e5784b8d6e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>