Add support for tonemapping
This change attempts to rectify many places where color spaces were
handled incorrectly. First off we should be doing all rendering in
linear color space. That means we need to convert any color data that is
in sRGB color space from input images to linear.  Once all rendering is
complete it is necessary to return to a color space supported by the
display, so several options are now available to tonemap the screen
colors.

Some other features that were made possible as well:
Skyboxes now use the same tonemapping as the rest of the scene.
It is now possible to use other HDR formats than RGBE (including RGBA32F
and RGBA16F).

Fixes: QTBUG-82335
Change-Id: I0965d6ed0efa6a93a926ccb3a5a24f2f06a52cb2
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>