Add additional built-in values for custom material frag shaders
This adds the following value for custom material fragment shaders:
vec3 NORMAL (renamed from WORLD_NORMAL)
vec3 TANGENT
vec3 BINORMAL
vec2 UV0 (readonly)
vec2 UV1 (readonly)

With these additional values, it is not possible to adjust normals such
that Normal Maps can be used in custom fragment shaders.  It is also no
longer necessary to manually create varyings for texture coordinates.

WORLD_NORMAL was renamed to NORMAL for consistency with TANGENT and
BINORMAL values. All are in the same coordinate space in the fragment
shader, and so the WORLD_ prefix is redudant.  We just document the
coordinate space instead.

For consistency in the fragment shader API, the WORLD_NORMAL argument
was renamed for all of the lighting methods as well because it
represents the same variable that is exposed in the custom MAIN.

UV0 and UV1 are now passed as varyings from the vertex shader. The code
was there before to set this up, but never got called because
CustomMaterials do not use the same path for image maps as
PrincipaledMaterials, so needed to be added manually.

Change-Id: Ifd2e9d631dd9a67be4cc5e65e1e304cdf4672f48
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>