Remove the getRenderContextInterface function
This approach was never very elegant and means we get a rather
unattractive connection to the Window, as clearly visible in the custom
material and effect code, where we needed to go through the Window to
get the RCI (render context interface). This change tries to make it
less painful to get access to the RCI by adding a pointer to it in
the scene manager.

Change-Id: I2cf9741d5dcbd934638d58e5cd726bc8653eb6ef
Reviewed-by: Andy Nichols <andy.nichols@qt.io>