Make POSITION optional in shaded custom materials
Previously in shaded custom materials vertex shaders it was required
that the POSITION property was set to pass the transformed vertex data
via gl_Position.

Now when there is a vertex shader that does not have POSITION set it is
assumed that we should just generate a default one based on the outputs
of the custom main function.

Change-Id: I46147c6ab9e59b7896ee6c5c4c5324a059a745b8
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>