Add support for PROJECTION_MATRIX and INVERSE_PROJECTION_MATRIX
It is now possible in Custom Material shaders to use the raw projection
and inverse projection matrix as uniforms. These uniforms are only
populated and exposed if they are used.

As part of this another uniform is necessary called NEAR_CLIP_VALUE
because when unprojecting depth values it is necessary to have depth
values be presented with respect to the near clip value being -1 or 0.

Change-Id: I48cb729e664e0ee5860caf558ab7ece67351b020
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>