Add custom material test for normal adjustments
This test checks the ability to adjust NORMAL, BINORMAL, and TANGENT
values in a custom material vertex shader, as well as use adjust the
NORMAL value by using a normal map in a fragment shader.

As a bonus this test checks the ability to read a texture from a vertex
shader.

Change-Id: I0f515b7b9ba1297b4a6ea38c8e63f7b49b72c08c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>