Adjust Light brightness to represent the real value
The brightness property of lights is a energy multiplier.  The default
value of 100 is nonsensical because we don't multiply the light energy
by 100, in fact we corrected this value by multiplying by 0.01 so a
value of 100 = a energy of 1.

There is no reason to keep this just because Qt 3D Studio used this
behavior.  It makes more sense to expose the raw energy value as
brightness, AKA a default brightness of 1.0

All examples/tests have been updated to reflect this.

In addition the documentation has been updated to reflect how light
attenuation is calculated using the properties of the SpotLight and
PointLight.

Change-Id: Ie03e47c5a14df0bb2e4617b5939a8c0d574665fa
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>