Unify effect shader style with custom materials
...as much as possible.

All existing effects are ported, with no API changes. The effect
system's rendering is not touched at all.

This changes the output slightly in effects that sample a texture
(that is not the input, screen texture) where there is < 1 alpha
pixels present, because it no longer (likely incorrectly)
premultiplies with the alpha. (the reason for doing this is lost
in time since Qt 3D Studio and is already marked as TODOs in the
RHI port, now we resolve it)
Typically this affects cases where there is a noise texture, e.g.
BrushStrokes or Scatter - it just sees slightly different noise now.
The other potentially affected effect is the bloom-tonemap since its
multi-pass structure (imho incorrectly) assumes pre-multiplied alpha
in the input that was output from a previous stage.

As a further inexplicable side effect, this fixes the macOS/Metal
Lancelot rendering for the following tests: aberration, depth, misc,
move. The baselines on the macOS machine are all broken for these atm
(something to do with the depth texture), but with this patch the runs
generate correct results.

Effect shaders work with the following built-in, special keywords.
The first set is identical to custom materials:
VARYING, FRAGCOLOR, CAMERA_PROPERTIES, FRAMEBUFFER_Y_UP,
NEAR_CLIP_VALUE, POSITION, VERTEX
The effect specifics are:
INPUT, INPUT_UV, TEXTURE_UV, INPUT_SIZE, OUTPUT_SIZE, FRAME, FPS

Besides these no built-in helper functions and macros are provided
(effect.glsllib is gone). No metadata block needs to be written by the
user since VARYING takes care of that internally. TEXTURE_UV is a
special case as that is based on a dynamically generated helper
function to flip the V on D3D/Vulkan/Metal, without making the user
write a condition based on FRAMEBUFFER_Y_UP. (like with SCREEN_TEXTURE
in custom materials, effects that sample a user provided texture
potentially need UV flipping; this does not apply for INPUT and
INPUT_UV since the MVP (which is otherwise just identity) does
invert Y as appropriate)

A vertex shader can also assign a custom value to POSITION but this is
rare. By default the standard gl_Position = mvp * attr_pos is added
automatically. qt_inputUV is passed through automatically as well, but
the vertex shader's custom main can modify INPUT_UV if it wants too.

The handling of textures is made compatible with custom materials: the
recently added filtering and mipmap generation features work identically
to custom materials now.

The list of supported property types is made compatible with custom
materials. (so added point, size, rect, matrix4x4, etc.)

The shader filenames are now provided as true URLs, not as a QByteArray.
The semantics are the same as with custom materials and Qt Quick source
URLs.

A new customeffect example is added that is focused on showing that
writing custom shaders for Effect is plain and simple.

The (non-existing) documentation for using Effect with custom shaders,
and Shader, Buffer, BufferInput, DepthInput, etc. docs are to be updated
separately in a follow-up commit.

The QUrl change is the only change in the shaderutils API. It should be
considered as a follow-up if removing DepthInput (DEPTH_TEXTURE could
fulfill the same role without needed any actual QML API), or sanitizing
some properties in BufferInput would make sense.

Change-Id: I4e286c8e92d26adfca9766b945189d9d2e1613e5
Reviewed-by: Andy Nichols <andy.nichols@qt.io>