Expose additional variables to custom shader lighting methods
For the light diffuse methods the following values were exposed:
METALNESS
ROUGHNESS
As these are required for some diffuse techniques.
For the specular light method the following values were exposed:
FRESNEL_CONTRIB
METALNESS
ROUGHNESS
SPECULAR_AMOUNT
The FRESNEL_CONTRIB value contains the built in fresnel contribution
that is done automatically and is needed if you want to use that
calculation (otherwise it's wasted when using a custom specular method).
It's also possible to continue ignoring this value.
METALNESS and ROUGHNESS are needed for certain specular rendering
techniques.  METALNESS is needed if the user wants to calculate their
own Fresnel. ROUGHNESS is used by our own built-in Blinn/Phong specular
calculation, so that should be and is used by other techniques as well.
SPECULAR_AMOUNT is needed in the case that the user wants to calculate
their own Fresnel value using the specular value provided by MAIN.

I also documented that BASE_COLOR and NORMAL is already exposed to all
of these lighting methods.

Change-Id: I7fb0dc3ec5b1f79ec90f91ac976f753c73cb8b41
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>