Make the PrincipledMaterial render as intended
The PrincipledMaterial previously was not rendering as intended, and
this patch intends to rectify many of those issues.

Diffuse BSDF:
There appears to only be a Lambert diffuse model in the default material
generator, however the intention was to use a version of the diffuse
model described in the paper by Brent Burley which describes the
methodology behind the design of the PrincipledMaterial. This change
will use the so-called Burley BSDF technique for the diffuse model when
using PrincipledMaterial.

Specular BSDF:
For the specular model the intended behavior was to always use GGX
specular shading. Unlike the DefaultMaterial we don't even offer the
ability to select a specular model, because we want to force GGX, but
what previously was happening is that we were falling back to a Blinn
specular model which leads to poor quality specular blobs.

Fresnel Calculations:
In default material Fresnel is calculated once globally and then applied
as a contribution to each specular calculation for each light in the
scene. This works as a fast approximation, but leads to poor Fresnel
quality when used in combination with PrincipledMaterial. So for the
PrincipledMaterial this global calculation is only used in combination
with IBL based specular calculations.  For the other light types the
fresnel calculation is part of individual lights specular calcuations.

Lancelot Tests:
New lancelot tests have been added to cover the rendering of
PrincipledMaterials with each lighting method.

PrincipledMaterial:
Any "smartness" on the frontend to account for metalness vs dielectric
is removed. On the frontend we just set properties and the backend
decides how to handle that. That means things like setting the
specularTint despite the material being metallic. The backend will do
the correct thing, so just pass it the data.  The default specularAmount
value has now been changed to 0.5 and the index of refraction property
has been removed. The index of refraction is determined by the specular
amount, and a default value of 0.5 will produce a IOR of 1.5 and a
dielectric normal incidence value of 0.04.  There is no need for an
explicit IOR property as long as we have a specularAmount property. All
calculations regarding luminance are gone, as this was only used to
correct for your lack of tonemapping, so it's no longer needed.

IBL:
The specular term for PrincipledMaterials when using IBL is now
calculated differntly. The roughness values used to look up which mip
map level is used for calculation are now "adjusted" to achieve a more
uniform look. This balances out the "roughness" of the specular
reflections more evenly for the middle roughnesses. Before items needed
to have a roughness greater than 0.6 before they stoped seeming really
glossy.

Materials:
We no longer bake the diffuse color of the material into each of the
lights diffuse channel. This might have been useful at some point in
time, but since then we have been multiplying lights by the diffuse
color anyway, which means we were applying this contribution twice,
leading to unusual behavior with certain colors. Expect lots of broken
lancelot tests...

Change-Id: Ia2993819f451b61cf6f8247f99eb1505b364cdea
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>