Avoid Vulkan validation errors when toggling effects
...by recreating the VkRenderpass when the texture format changes.

Effects like to default to RGBA32F for their textures. This means
however that toggling between having some effects and having no effects
means a change in the format of the View3D's backing QRhiTexture. With
Vulkan the renderpass depends on the texture format. Therefore it's not
enough to to a create() on the QRhiTextureRenderTarget, it also needs a
new QRhiRenderPassDescriptor.

The same goes in the effect system, when multi-pass effect pick an
unused texture for the Buffer objects. If reusing an existing texture,
if the format is different the rendertarget needs a new renderpass.

Change-Id: Ide9c217a8fa3ffcc1e094b2fff6adceb91dddbce
Reviewed-by: Andy Nichols <andy.nichols@qt.io>