Fix toggling effects when MSAA or SSAA is on
Toggling MSAA/SSAA when there is at least one effect on the View3D
works. The opposite fails: adding the first effect after MSAA or SSSA
has been enabled leads to a validation failure with Vulkan and an
actual error with D3D. This is due to the m_texture's format changing
between RGBA8 and RGBA32F, which then should be reflected in the MSAA
renderbuffer and SSAA texture as well but the existing logic did not
allow for this.

Change-Id: I7752681169815338c77ab3d3e86f9b35fe49dad8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>