Remove Support for AreaLight
AreaLights while part of the API are broken in much of the existing
rendering paths. We could fix this, but the techniques used to render
real-time AreaLights are very expensive, and don't really make sense for
the hardware platforms we see as our targets (Embedded).  So instead we
should remove them for 6.0 with the possibility of re-adding support
later if there is an actual demand.

Previously in Qt 3D Studio (and Qt Quick 3D by extension) we had support
for AreaLights, but not spot lights, and so AreaLights was the only way
to have a directional light with falloff.  Since then we have added
support for SpotLights which should fulfill this use case without
compromising runtime performance.

Fixes: QTBUG-86488
Change-Id: Ic0b9e5c46319def09c52a8766db306a9ca95724a
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>