Make input, output, uniform lists sorted
We already do this for uniforms. Now extend it to all the relevant maps.
This helps generating the exact same shader string on every run. Right
now the order of inputs and outputs is random (depends on what order
QHash decides to iterate), which also means that the input/output
locations can be different between runs. This is not fatal (or rather,
it was fatal with OpenGL until qtbase got fixed to accommodate for this),
but both in-memory and disk caches for shaders suffer if the exact same
material generates a different shader string in a different run of the
application just because QHash iterates in a different order now.

Task-number: QTBUG-86531
Change-Id: I2f6471de71576d220cdfd876a950b72accd16eda
Reviewed-by: Andy Nichols <andy.nichols@qt.io>