Commit texture operations on the QSGTexture
Consider the case when we get a QSGTexture that is not a
QSGDynamicTexture (i.e. something other than QSGLayer). In this case
nothing ensures that the texture is created and the data upload is
triggered. With Qt Quick calling commitTextureOperations() is done by
the QSGMaterial implementations. For Quick 3D purposes, it is up to the
buffer manager to do it before accessing the QRhiTexture.

Task-number: QTBUG-86530
Change-Id: Ied0ca3bd40ad3a673d5edd308ac7108ed56c0afd
Reviewed-by: Andy Nichols <andy.nichols@qt.io>