Fix object parenting in sourceItem Lancelot test
One cannot pass in a random object as the parent to createObject(). It
needs to be something an Item can normally be parented to (like some
other visual Item).

Attempting to parent to a Texture for example leads to:
QQmlComponent: Created graphical object was not placed in the graphics scene

Task-number: QTBUG-86530
Change-Id: I3685ee57040e799d2885be99a9088d33dba88cc6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>