Make sure QColors are converted to linear color space
Anytime we use QColor on the frontend we need to make sure that we
convert it to linear color space before using it as a uniform.  This
saves us from having to do the conversion in fragment shaders like we do
for textures that provide color data. This also makes sure that the
final output color is the intended color when things are brough back
into SRGB color space.

Change-Id: Ib8acedd904e2b8b27efa3b8c57ab67c11a002ff7
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>