Make Texture.sourceItem crash less
As we are giving up to create a strictly correct solution, fix up the
existing code in a way that makes the manual test fairly usable. Many
crashes that occur when changing the source item are now eliminated.

It is definitely not correct from thread synchronization/safety
perspective, and there's still odd things around those updateTexture()
calls. We are still unable to get content in the first frame.

However, this is the most stable approach we ever had. The manual test
feels very usable and stable in fact.

Task-number: QTBUG-86530
Change-Id: I648c5600b00a830c4a8ae15a807ea073a8c24bc0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>