Purge the env vars page
The standard env vars to control the rhi backend are now described on
the graphics page. The remaining two that are on the current env var
page may not even be functional. So remove this page altogether. If
there is a need to document env vars it can be done in the other pages.

Fixes: QTBUG-85963
Change-Id: I27ba6d40e3486c4885aac4ec30f4934dba4cc7d4
Reviewed-by: Christian Strømme <christian.stromme@qt.io>