Avoid losing and recalculating shader-related data between frames
QSSGRhiShaderStagesWithResources is not actually persistent. It might
have been in the original RHI port, but now it is often transient, and
is lost in the next frame (similarly to layer prep data et al). This is
fine but we should not be doing work over and over again.

Move persistent data to QSSGRhiShaderStages.

Also avoid various unnecessary lookups, especially for uniforms that are
set from the material but not present in the actual shaders (these were
looked up with an expensive loop again and again).

And even where lookups are needed for uniform block members and vertex inputs,
make QSSGRhiShaderStages provide a hash map. (iterating QShaderDescription
lists and looking for an (int) binding is probably fine, but bytearray
comparisons based on name is not, esp. when it's a uniform block member
list with > 10 entries) Do the same for vertex inputs as well.

Also fixes sub-optimal handling of m_uniformArrays (no need to 'new' the
members of that list).

Changes also how vertex input locations are calculated.
QSSGRhiInputAssemblerState's basic caching mechanism is useless in
practice (due to how an ia object used over its lifetime) so that's now
removed. Instead the lookup is made faster.

Change-Id: I7af8c39589e1e2f8ac0470acbcd013fd855fddab
Reviewed-by: Andy Nichols <andy.nichols@qt.io>