Improve the shader key string visitor
The shader key property names were never intended to be used for
anything but debugging purposes, unfortunately we've come to rely on it
quite a bit in Qt6. While the real solution is to stop using the
shader key string as a key for cache look-up, the string visitor and
the surrounding shader key code was in great need of some clean-up
to avoid unnecessary string conversions, char literals as string etc.
This change also tries to avoid any extra allocations during string
concatenations by reserving enough data in the buffer to contain all
the property names (and then some).

Change-Id: Ifb6c9b5c553d8576739b041e0e82f955b9314bb9
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>