Use cubeMaps for IBL texture sources
This changes how IBL sources are pre-processed and used at runtime.
Previously we would directly use the equirectangular texture data. This
led to several issues:
- A Compute shader was required at runtime to pre-filter each mip level
(which depended on the previous level).  Compute shaders are not
available everywhere, and doing the software version at runtime was very
expensive.
- Even when we had support for compute shaders the minimium number of
compute units available was not enough for decent processing.
- The method we were using for pre-filtering was not what we needed for
doing proper BRDF.
- Using equirectangular textures at runtime is expensive because each
time we want to sample the texture we have to convert the coordinate
space.
- Using equirectangular projections leads to artifacts at the poles
which lead to obvious artifacts in the skybox, as well as specular
reflections which depend on values samples against those poles.
- The lack of irradiance data let to very dark diffuse contributions.

So instead of using a prefiltered equirectangular texture directly, now
we create a cubeMap from the provided texture using a series of standard
shading passes which contains pre-filtered environment mip levels for
BRDF as well as a pre-convolved irradiance map. This eliminates all of
the above issues.

Fixes: QTBUG-78849
Fixes: QTBUG-86682
Fixes: QTBUG-86684
Fixes: QTBUG-86685
Fixes: QTBUG-86795
Change-Id: I7dcd059c8b669142fc5c368f4dbfeed199a57f16
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>